clean
store/index.jsの分割
