+++ src/App.vue
@@ -28,7 +28,7 @@ export default {
   },
   methods: {
     logout() {
-      this.$store.dispatch('logout');
+      this.$store.dispatch('logout/logout');
     }
   },
 }
