+++ src/components/pages/Login.vue
@@ -30,7 +30,7 @@ export default {
   },
   methods: {
     login(){
-      this.$store.dispatch('login', {
+      this.$store.dispatch('login/login', {
         email: this.email,
         password: this.password
       })
