+++ src/store/modules/logout.js
@@ -1,0 +1,16 @@
+import router from "../../router";
+
+const actions = {
+  logout({commit}) {
+    commit('updateIdToken', null, {root: true});
+    localStorage.removeItem('idToken');
+    localStorage.removeItem('refreshToken');
+    localStorage.removeItem('expiryTimeMs');
+    router.replace('/login');
+  },
+};
+
+export default {
+  namespaced: true,
+  actions
+};
